wswan: merge drivers/wswan.cpp, includes/wswan.h, and machine/wswan.cpp (nw)
